+++ include/render/Integrator.h
@@ -30,6 +30,8 @@ namespace render {
     std::vector<std::uint64_t> activeSamplesPerDepth;
     std::vector<std::uint64_t> frontierRayHitsPerDepth;
     std::vector<std::uint64_t> frontierRayMissesPerDepth;
+    std::vector<std::uint64_t> frontierPacketChunksPerDepth;
+    std::vector<std::uint64_t> frontierScalarRaysPerDepth;
     std::uint64_t activeSampleDepthsProcessed{0};
     std::vector<double> radianceDeltaSquaredSumPerDepth;
     std::vector<double> maxRadianceDeltaPerDepth;
@@ -42,6 +44,7 @@ namespace render {
     void reset(bool scalarFallback);
     void recordActiveDepth(std::uint64_t activeSamples);
     void recordFrontierIntersections(std::uint64_t hitRays, std::uint64_t missRays);
+    void recordFrontierTraversal(std::uint64_t packetChunks, std::uint64_t scalarRays);
     void recordRadianceDeltaDepth(double squaredSum, double maxDelta);
   };